Refactor code, move filter to widget\editor.py
