+++ spyderlib/widgets/tabs.py
@@ -13,7 +13,7 @@
 
 from spyderlib.qt.QtGui import (QTabWidget, QMenu, QDrag, QApplication,
                                 QTabBar, QWidget, QHBoxLayout)
-from spyderlib.qt.QtCore import Signal, Qt, QPoint, QMimeData, QByteArray, QObject, QEvent
+from spyderlib.qt.QtCore import Signal, Qt, QPoint, QMimeData, QByteArray
 
 import os.path as osp
 
@@ -101,29 +101,6 @@ class TabBar(QTabBar):
             self.sig_move_tab.emit(index_from, index_to)
             event.acceptProposedAction()
         QTabBar.dropEvent(self, event)
-        
-
-class TabFilter(QObject):
-    """
-    This event filter is installed on the QTabBar of BaseTabs to allow for
-    reordering of tabs in the editorstacks.
-    """
-    def __init__(self, editor_stack):
-        QObject.__init__(self)
-        self.editor_stack = editor_stack
-
-    def eventFilter(self,  obj,  event):
-        event_type = event.type()
-        if event_type == QEvent.MouseButtonPress:
-            self.editor_stack.tab_pressed(event)
-            return True
-        if event_type == QEvent.MouseMove:
-            self.editor_stack.tab_moved(event)
-            return True
-        if event_type == QEvent.MouseButtonRelease:
-            self.editor_stack.tab_released(event)
-            return True
-        return False
 
         
 class BaseTabs(QTabWidget):
@@ -135,8 +112,6 @@ class BaseTabs(QTabWidget):
         QTabWidget.__init__(self, parent)
         
         self.setUsesScrollButtons(True)
-        self.filter = TabFilter(parent)  # Need to keep reference
-        self.tabBar().installEventFilter(self.filter)
         self.corner_widgets = {}
         self.menu_use_tooltips = menu_use_tooltips
